refactor: split WebAuthn server logic into separate files
